KyboardHook 関連の動作をBehavior化
